Updating requirements to stay in line with analytics-data-api
Upgrade to DRF removed the default UTC formatting for serialization:
https://github.com/encode/django-rest-framework/blob/master/docs/community/release-notes.md#370

This resulted in our dates being set to local time. We now set UTC as the default timezone in
settings/test.py. This required updating the other (non serializer) tests that were expecting
a timezone offset and had their expected values hardcoded to the fixture +6 hours.