ENT-1336. Handle case when pgp_key is empty string
The pgp_key config in the LMS is a TextField which means when it is
updated the key is set to empty string ('') and not None. We were only
checking for None and therefore passing along an empty pgp_key to the
_get_encrypted_file method. This is fixed with this PR by doing a truthy
check - of which both empty string and None will fail.