Remove pip-tools from base requirements and upgrade requirements (#165)
See https://openedx.atlassian.net/projects/BOM/issues/BOM-1359.

This error message is thrown while upgrading edx-analytics-data-api: edx-enterprise-data 1.3.13 has requirement pip-tools==4.3.0, but you'll have pip-tools 4.5.1 which is incompatible. See: https://github.com/edx/edx-analytics-data-api/pull/299

Since pip-tools is not an actual requirement of edx-enterprise-data I am removing that from base.in and upgrading the requirements. Also it was removed tox from base.in requirements.

Also it was added a constraints.txt file to accomplish the OEP-18 practices and a pip-tools.in file to keep the version of the tools used to upgrade the dependencies.

Constraints changed:

Added:
edx-django-utils<3.0.0
edx-rest-api-client<4.0.0

tox-battery from 0.5.1 to 0.5.2

tox from 3.10.0 to none