Fix tox (#183)
Currently, tox is not running the actual commands for each environment (data and reporting). This is due to an error from this PR https://github.com/edx/edx-enterprise-data/pull/179
This solves that.

I am removing [testenv:data] and [testenv:reporting] sections in order to avoid the creation of a new section for each django version.