fix: add rate-limiting retry for enterprise requests
MST-1304. Everytime the enterprise-reporting jenkins job runs, it creates spikes of errors for prod analytics api. This is because the jenkins job makes many calls to an endpoint that is rate limited, and eventually errors out when that rate limit is reached. If we know that we may get a rate limiting error, we should catch that exception and retry after the appropriate amount of time has passed.